+++ Assets/Adjust/Editor/AdjustEditor.cs
@@ -12,21 +12,21 @@ using UnityEditor.Callbacks;
 using UnityEditor.iOS.Xcode;
 #endif
 
-public class AdjustEditor
+public class AdjustEditor : AssetPostprocessor
 {
     private static bool isPostProcessingEnabled = true;
 
-    [MenuItem("Assets/Adjust/Check post processing status")]
+    [MenuItem("Assets/Adjust/Check Post Processing Permission")]
     public static void CheckPostProcessingPermission()
     {
-        EditorUtility.DisplayDialog("[Adjust]", "The post processing for Adjust SDK is " + (isPostProcessingEnabled ? "enabled." : "disabled."), "OK");
+        EditorUtility.DisplayDialog("Adjust SDK", "The post processing for Adjust SDK is " + (isPostProcessingEnabled ? "enabled." : "disabled."), "OK");
     }
 
-    [MenuItem("Assets/Adjust/Change post processing status")]
+    [MenuItem("Assets/Adjust/Change Post Processing Permission")]
     public static void ChangePostProcessingPermission()
     {
         isPostProcessingEnabled = !isPostProcessingEnabled;
-        EditorUtility.DisplayDialog("[Adjust]", "The post processing for Adjust SDK is now " + (isPostProcessingEnabled ? "enabled." : "disabled."), "OK");
+        EditorUtility.DisplayDialog("Adjust SDK", "The post processing for Adjust SDK is now " + (isPostProcessingEnabled ? "enabled." : "disabled."), "OK");
     }
 
     [MenuItem("Assets/Adjust/Export Unity Package")]
@@ -36,16 +36,14 @@ public class AdjustEditor
         string assetsPath = "Assets/Adjust";
         List<string> assetsToExport = new List<string>();
 
-        // Adjust Editor script.
-        assetsToExport.Add("Assets/Editor/AdjustEditor.cs");
         // Adjust Assets.
-        assetsToExport.Add(assetsPath + "/Adjust.cs");
-        assetsToExport.Add(assetsPath + "/Adjust.prefab");
         assetsToExport.Add(assetsPath + "/3rd Party/SimpleJSON.cs");
 
         assetsToExport.Add(assetsPath + "/Android/adjust-android.jar");
         assetsToExport.Add(assetsPath + "/Android/AdjustAndroid.cs");
         assetsToExport.Add(assetsPath + "/Android/AdjustAndroidManifest.xml");
+
+        assetsToExport.Add(assetsPath + "/Editor/AdjustEditor.cs");
 
         assetsToExport.Add(assetsPath + "/ExampleGUI/ExampleGUI.cs");
         assetsToExport.Add(assetsPath + "/ExampleGUI/ExampleGUI.prefab");
@@ -67,6 +65,9 @@ public class AdjustEditor
         assetsToExport.Add(assetsPath + "/iOS/AdjustUnityDelegate.h");
         assetsToExport.Add(assetsPath + "/iOS/AdjustUnityDelegate.mm");
 
+        assetsToExport.Add(assetsPath + "/Prefab/Adjust.prefab");
+
+        assetsToExport.Add(assetsPath + "/Unity/Adjust.cs");
         assetsToExport.Add(assetsPath + "/Unity/AdjustAttribution.cs");
         assetsToExport.Add(assetsPath + "/Unity/AdjustConfig.cs");
         assetsToExport.Add(assetsPath + "/Unity/AdjustEnvironment.cs");