+++ Assets/Adjust/iOS/AdjustiOS.cs
@@ -8,7 +8,7 @@ namespace com.adjust.sdk
 #if UNITY_IOS
     public class AdjustiOS
     {
-        private const string sdkPrefix = "unity4.19.1";
+        private const string sdkPrefix = "unity4.19.2";
 
         [DllImport("__Internal")]
         private static extern void _AdjustLaunchApp(